core: replace project files
