+++ Blasphemous.Randomizer.Multiworld.sln
@@ -1,0 +1,25 @@
+﻿
+Microsoft Visual Studio Solution File, Format Version 12.00
+# Visual Studio Version 17
+VisualStudioVersion = 17.8.34408.163
+MinimumVisualStudioVersion = 10.0.40219.1
+Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "Blasphemous.Randomizer.Multiworld", "Blasphemous.Randomizer.Multiworld\Blasphemous.Randomizer.Multiworld.csproj", "{D2878996-992F-49A6-8612-6AB3B18D375D}"
+EndProject
+Global
+	GlobalSection(SolutionConfigurationPlatforms) = preSolution
+		Debug|Any CPU = Debug|Any CPU
+		Release|Any CPU = Release|Any CPU
+	EndGlobalSection
+	GlobalSection(ProjectConfigurationPlatforms) = postSolution
+		{D2878996-992F-49A6-8612-6AB3B18D375D}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
+		{D2878996-992F-49A6-8612-6AB3B18D375D}.Debug|Any CPU.Build.0 = Debug|Any CPU
+		{D2878996-992F-49A6-8612-6AB3B18D375D}.Release|Any CPU.ActiveCfg = Release|Any CPU
+		{D2878996-992F-49A6-8612-6AB3B18D375D}.Release|Any CPU.Build.0 = Release|Any CPU
+	EndGlobalSection
+	GlobalSection(SolutionProperties) = preSolution
+		HideSolutionNode = FALSE
+	EndGlobalSection
+	GlobalSection(ExtensibilityGlobals) = postSolution
+		SolutionGuid = {F4764644-136A-4C4C-987F-5DE44120D6F5}
+	EndGlobalSection
+EndGlobal
