+++ Blasphemous.Randomizer.Multiworld/AP/ArchipelagoLocation.cs
@@ -1,5 +1,5 @@
 ﻿
-namespace BlasphemousMultiworld.AP
+namespace Blasphemous.Randomizer.Multiworld.AP
 {
     public class ArchipelagoLocation
     {
