+++ Blasphemous.Randomizer.Multiworld/AP/Receivers/HintReceiver.cs
@@ -2,7 +2,7 @@
 using Framework.Managers;
 using System.Collections.Generic;
 
-namespace BlasphemousMultiworld.AP.Receivers
+namespace Blasphemous.Randomizer.Multiworld.AP.Receivers
 {
     public class HintReceiver
     {
