+++ Blasphemous.Randomizer.Multiworld/GameSettings.cs
@@ -1,6 +1,5 @@
-﻿using BlasphemousRandomizer;
-
-namespace BlasphemousMultiworld
+﻿
+namespace Blasphemous.Randomizer.Multiworld
 {
     public class GameSettings
     {
