+++ Blasphemous.Randomizer.Multiworld/MultiworldPersistenceData.cs
@@ -1,11 +1,11 @@
-﻿using System;
+﻿using Blasphemous.ModdingAPI.Persistence;
+using System;
 using System.Collections.Generic;
-using ModdingAPI;
 
-namespace BlasphemousMultiworld
+namespace Blasphemous.Randomizer.Multiworld
 {
     [Serializable]
-    public class MultiworldPersistenceData : ModPersistentData
+    public class MultiworldPersistenceData : SaveData
     {
         public MultiworldPersistenceData() : base("ID_MULTIWORLD") { }
