+++ Blasphemous.Randomizer.Multiworld/QueuedItem.cs
@@ -1,0 +1,16 @@
+﻿
+namespace Blasphemous.Randomizer.Multiworld;
+
+public class QueuedItem
+{
+    public QueuedItem(string itemId, int index, string player)
+    {
+        this.itemId = itemId;
+        this.index = index;
+        this.player = player;
+    }
+
+    public string itemId;
+    public int index;
+    public string player;
+}
