--- BlasphemousMultiworld.sln
@@ -1,25 +1,0 @@
-﻿
-Microsoft Visual Studio Solution File, Format Version 12.00
-# Visual Studio Version 16
-VisualStudioVersion = 16.0.31605.320
-MinimumVisualStudioVersion = 10.0.40219.1
-Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "BlasphemousMultiworld", "BlasphemousMultiworld\BlasphemousMultiworld.csproj", "{2ADF79F6-AFA8-4FB0-8C32-95CC93D98DD6}"
-EndProject
-Global
-	GlobalSection(SolutionConfigurationPlatforms) = preSolution
-		Debug|Any CPU = Debug|Any CPU
-		Release|Any CPU = Release|Any CPU
-	EndGlobalSection
-	GlobalSection(ProjectConfigurationPlatforms) = postSolution
-		{2ADF79F6-AFA8-4FB0-8C32-95CC93D98DD6}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
-		{2ADF79F6-AFA8-4FB0-8C32-95CC93D98DD6}.Debug|Any CPU.Build.0 = Debug|Any CPU
-		{2ADF79F6-AFA8-4FB0-8C32-95CC93D98DD6}.Release|Any CPU.ActiveCfg = Release|Any CPU
-		{2ADF79F6-AFA8-4FB0-8C32-95CC93D98DD6}.Release|Any CPU.Build.0 = Release|Any CPU
-	EndGlobalSection
-	GlobalSection(SolutionProperties) = preSolution
-		HideSolutionNode = FALSE
-	EndGlobalSection
-	GlobalSection(ExtensibilityGlobals) = postSolution
-		SolutionGuid = {2EADD280-4234-4962-A8F1-20CD1FD9069D}
-	EndGlobalSection
-EndGlobal
