--- BlasphemousMultiworld/BlasphemousMultiworld.csproj
@@ -1,77 +1,0 @@
-﻿<Project Sdk="Microsoft.NET.Sdk">
-
-  <PropertyGroup>
-    <TargetFramework>net35</TargetFramework>
-    <AssemblyName>BlasphemousMultiworld</AssemblyName>
-    <BepInExPluginGuid>com.damocles.blasphemous.multiworld</BepInExPluginGuid>
-    <BepInExPluginName>Multiworld</BepInExPluginName>
-    <Description>A multiworld client for the Blasphemous randomizer</Description>
-    <Version>2.1.3</Version>
-    <AllowUnsafeBlocks>true</AllowUnsafeBlocks>
-    <LangVersion>latest</LangVersion>
-  </PropertyGroup>
-
-  <ItemGroup>
-    <PackageReference Include="Archipelago.MultiClient.Net" Version="5.0.4" />
-    <PackageReference Include="BepInEx.Analyzers" Version="1.*" PrivateAssets="all" />
-    <PackageReference Include="BepInEx.Core" Version="5.*" />
-    <PackageReference Include="BepInEx.PluginInfoProps" Version="1.*" />
-    <PackageReference Include="Blasphemous.ModdingReferences" Version="4.0.67" />
-    <PackageReference Include="UnityEngine.Modules" Version="2017.4.40" IncludeAssets="compile" />
-  </ItemGroup>
-  
-  <ItemGroup Condition="'$(TargetFramework.TrimEnd(`0123456789`))' == 'net'">
-    <PackageReference Include="Microsoft.NETFramework.ReferenceAssemblies" Version="1.0.2" PrivateAssets="all" />
-  </ItemGroup>
-  
-  <ItemGroup>
-    <Reference Include="BlasphemousRandomizer">
-      <HintPath>..\lib\BlasphemousRandomizer.dll</HintPath>
-    </Reference>
-    <Reference Include="ModdingAPI">
-      <HintPath>..\lib\ModdingAPI.dll</HintPath>
-    </Reference>
-  </ItemGroup>
-
-  <Target Name="Development" AfterTargets="PostBuildEvent">
-
-    <!-- Get resource files -->
-    <PropertyGroup>
-      <ResourcesFolder>$(SolutionDir)resources\</ResourcesFolder>
-    </PropertyGroup>
-    <ItemGroup>
-      <Config Include="$(ResourcesFolder)config\*" />
-      <DataDlls Include="$(ResourcesFolder)data\*" />
-      <DataFiles Include="$(ResourcesFolder)data\$(BepInExPluginName)\*" />
-      <Keybindings Include="$(ResourcesFolder)keybindings\*" />
-      <Levels Include="$(ResourcesFolder)levels\$(BepInExPluginName)\*" />
-      <Localization Include="$(ResourcesFolder)localization\*" />
-    </ItemGroup>
-
-    <!-- Copy plugin & resources to development folder -->
-    <PropertyGroup>
-      <DevFolder>C:\Program Files (x86)\Steam\steamapps\common\Blasphemous\Modding\</DevFolder>
-    </PropertyGroup>
-    <Copy SourceFiles="$(TargetPath)" DestinationFolder="$(DevFolder)plugins" />
-    <Copy SourceFiles="@(Config)" DestinationFolder="$(DevFolder)config" />
-    <Copy SourceFiles="@(DataDlls)" DestinationFolder="$(DevFolder)data" />
-    <Copy SourceFiles="@(DataFiles)" DestinationFolder="$(DevFolder)data\$(BepInExPluginName)" />
-    <Copy SourceFiles="@(Keybindings)" DestinationFolder="$(DevFolder)keybindings" />
-    <Copy SourceFiles="@(Levels)" DestinationFolder="$(DevFolder)levels\$(BepInExPluginName)" />
-    <Copy SourceFiles="@(Localization)" DestinationFolder="$(DevFolder)localization" />
-
-    <!-- Export & zip all mod files -->
-    <PropertyGroup>
-      <ExportFolder>$(SolutionDir)export\$(TargetName)\</ExportFolder>
-    </PropertyGroup>
-    <Copy SourceFiles="$(TargetPath)" DestinationFolder="$(ExportFolder)plugins" />
-    <Copy SourceFiles="@(Config)" DestinationFolder="$(ExportFolder)config" />
-    <Copy SourceFiles="@(DataDlls)" DestinationFolder="$(ExportFolder)data" />
-    <Copy SourceFiles="@(DataFiles)" DestinationFolder="$(ExportFolder)data\$(BepInExPluginName)" />
-    <Copy SourceFiles="@(Keybindings)" DestinationFolder="$(ExportFolder)keybindings" />
-    <Copy SourceFiles="@(Levels)" DestinationFolder="$(ExportFolder)levels\$(BepInExPluginName)" />
-    <Copy SourceFiles="@(Localization)" DestinationFolder="$(ExportFolder)localization" />
-    <ZipDirectory SourceDirectory="$(ExportFolder)" DestinationFile="$(ExportFolder)..\$(TargetName).zip" Overwrite="true" />
-
-  </Target>
-</Project>
