--- BlasphemousMultiworld/Main.cs
@@ -1,21 +1,0 @@
-﻿using BepInEx;
-using BlasphemousRandomizer;
-
-namespace BlasphemousMultiworld
-{
-    [BepInPlugin(PluginInfo.PLUGIN_GUID, PluginInfo.PLUGIN_NAME, PluginInfo.PLUGIN_VERSION)]
-    [BepInDependency("com.damocles.blasphemous.modding-api", "1.3.4")]
-    [BepInDependency("com.damocles.blasphemous.randomizer", "2.0.0")]
-    [BepInProcess("Blasphemous.exe")]
-    public class Main : BaseUnityPlugin
-    {
-        public static Multiworld Multiworld { get; private set; }
-        public static Randomizer Randomizer { get; private set; }
-
-        private void Start()
-        {
-            Randomizer = BlasphemousRandomizer.Main.Randomizer;
-            Multiworld = new Multiworld(PluginInfo.PLUGIN_GUID, PluginInfo.PLUGIN_NAME, PluginInfo.PLUGIN_VERSION);
-        }
-    }
-}
